+++ src/UnityFx.Async.AssetStore/Assets/Plugins/UnityFx.Async/Scripts/Helpers/WwwResult{T}.cs
@@ -68,7 +68,7 @@ namespace UnityFx.Async.Helpers
 		/// </summary>
 		protected virtual T GetResult(WWW request)
 		{
-			return AsyncUtility.GetResult<T>(request);
+			return AsyncWww.GetResult<T>(request);
 		}
 
 		#endregion