+++ src/UnityFx.Async/Implementation/Core/MultiContextCallbackCollection.cs
@@ -9,25 +9,13 @@ using System.Threading;
 namespace UnityFx.Async
 {
 	/// <summary>
-	/// A default implementation of <see cref="IAsyncCallbackCollection"/>. The implementation assumes that in most cases
+	/// An implementation of <see cref="IAsyncCallbackCollection"/>. The implementation assumes that in most cases
 	/// there are 3 or less completion callbacks and 1 or less progress callbacks. A <see cref="SynchronizationContext"/>
 	/// instance is stored for each callback.
 	/// </summary>
-	internal class DefaultCallbackCollection : IAsyncCallbackCollection
+	internal class MultiContextCallbackCollection : IAsyncCallbackCollection
 	{
 		#region data
-
-		private struct CallbackData
-		{
-			public readonly object Callback;
-			public readonly SynchronizationContext SyncContext;
-
-			public CallbackData(object callback, SynchronizationContext syncContext)
-			{
-				Callback = callback;
-				SyncContext = syncContext;
-			}
-		}
 
 		private readonly IAsyncOperation _op;
 
@@ -43,15 +31,9 @@ namespace UnityFx.Async
 
 		#region interface
 
-		public DefaultCallbackCollection(IAsyncOperation op)
+		public MultiContextCallbackCollection(IAsyncOperation op)
 		{
 			_op = op;
-		}
-
-		public DefaultCallbackCollection(IAsyncOperation op, object callback, SynchronizationContext syncContext)
-		{
-			_op = op;
-			_completionCallback1 = new CallbackData(callback, syncContext);
 		}
 
 		#endregion
@@ -175,37 +157,37 @@ namespace UnityFx.Async
 		{
 			if (_progressCallback1.Callback != null)
 			{
-				InvokeProgressCallback(_progressCallback1);
+				CallbackUtility.InvokeProgressCallback(_op, _progressCallback1.Callback, _progressCallback1.SyncContext);
 			}
 
 			if (_progressCallbacks != null)
 			{
 				foreach (var item in _progressCallbacks)
 				{
-					InvokeProgressCallback(item);
+					CallbackUtility.InvokeProgressCallback(_op, item.Callback, item.SyncContext);
 				}
 			}
 
 			if (_completionCallback1.Callback != null)
 			{
-				Invoke(_completionCallback1, invokeAsync);
+				CallbackUtility.InvokeCompletionCallback(_op, _completionCallback1.Callback, _completionCallback1.SyncContext, invokeAsync);
 			}
 
 			if (_completionCallback2.Callback != null)
 			{
-				Invoke(_completionCallback2, invokeAsync);
+				CallbackUtility.InvokeCompletionCallback(_op, _completionCallback2.Callback, _completionCallback2.SyncContext, invokeAsync);
 			}
 
 			if (_completionCallback3.Callback != null)
 			{
-				Invoke(_completionCallback3, invokeAsync);
+				CallbackUtility.InvokeCompletionCallback(_op, _completionCallback3.Callback, _completionCallback3.SyncContext, invokeAsync);
 			}
 
 			if (_completionCallbacks != null)
 			{
 				foreach (var item in _completionCallbacks)
 				{
-					Invoke(item, invokeAsync);
+					CallbackUtility.InvokeCompletionCallback(_op, item.Callback, item.SyncContext, invokeAsync);
 				}
 			}
 		}
@@ -214,14 +196,14 @@ namespace UnityFx.Async
 		{
 			if (_progressCallback1.Callback != null)
 			{
-				InvokeProgressCallback(_progressCallback1);
+				CallbackUtility.InvokeProgressCallback(_op, _progressCallback1.Callback, _progressCallback1.SyncContext);
 			}
 
 			if (_progressCallbacks != null)
 			{
 				foreach (var item in _progressCallbacks)
 				{
-					InvokeProgressCallback(item);
+					CallbackUtility.InvokeProgressCallback(_op, item.Callback, item.SyncContext);
 				}
 			}
 		}
@@ -229,67 +211,6 @@ namespace UnityFx.Async
 		#endregion
 
 		#region implementation
-
-		private void Invoke(CallbackData callbackData, bool invokeAsync)
-		{
-			var syncContext = callbackData.SyncContext;
-
-			if (invokeAsync)
-			{
-				if (syncContext != null)
-				{
-					syncContext.Post(InvokeInline, callbackData.Callback);
-				}
-				else
-				{
-					syncContext = SynchronizationContext.Current;
-
-					if (syncContext != null)
-					{
-						syncContext.Post(InvokeInline, callbackData.Callback);
-					}
-					else
-					{
-						ThreadPool.QueueUserWorkItem(InvokeInline, callbackData.Callback);
-					}
-				}
-			}
-			else if (syncContext == null || syncContext == SynchronizationContext.Current)
-			{
-				InvokeInline(callbackData.Callback);
-			}
-			else
-			{
-				syncContext.Post(InvokeInline, callbackData.Callback);
-			}
-		}
-
-		private void InvokeProgressCallback(CallbackData callbackData)
-		{
-			var syncContext = callbackData.SyncContext;
-
-			if (syncContext == null || syncContext == SynchronizationContext.Current)
-			{
-				InvokeProgressChangedInline(callbackData.Callback);
-			}
-			else
-			{
-				syncContext.Post(InvokeProgressChangedInline, callbackData.Callback);
-			}
-		}
-
-		private void InvokeInline(object callback)
-		{
-			Debug.Assert(callback != null);
-			CallbackUtility.InvokeCompletionCallback(_op, callback);
-		}
-
-		private void InvokeProgressChangedInline(object callback)
-		{
-			Debug.Assert(callback != null);
-			CallbackUtility.InvokeProgressCallback(_op, callback);
-		}
-
 		#endregion
 	}
 }