+++ CHANGELOG.md
@@ -8,13 +8,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/); this proj
 
 ### Added
 - Added `Play`/`Wait` extension methods for `Animation` and `Animator`.
-- Added `AsyncCreationOptions.UseSharedSynchronizationContext` creation option that allows using a shared synchronization context for the operation continuations.
-- Added `AsyncResult.DefaultCreationOptions` static property to set options used by all operations.
-- Added `AsyncResult.SharedSynchronizationContext` static property to be used by continuations if `AsyncCreationOptions.UseSharedSynchronizationContext` option is set.
+- Added `AsyncResult.IsStarted` helper property.
 
 ### Changed
 - Changed `AsyncResult` constructors argument order to avoid ambiguity in some cases.
 - Moved the package content to Plugins folder and remove assembly definition file.
+- Moved web request related helpers from `AsyncUtility` to `AsyncWww` class.
+- Changed `AsyncUtility.SendToMainThread`, `AsyncUtility.PostToMainThread` and `AsyncUtility.InvokeOnMainThread` implementation to use `ConcurrentQueue` for net46+ to avoid unnesesary locks.
+- Changed interface of `AsyncResultQueue`.
+
+### Removed
+- Removed `IAsyncOperationEvents.TryAddCompletionCallback` and `IAsyncOperationEvents.TryAddProgressCallback` methods. These methods are not needed in 99% of cases and may lead to logic errors in multi-threaded environment.
 
 -----------------------
 ## [0.9.5] - 2018.07.31